+++ cmd/main.go
@@ -216,5 +216,4 @@ func main() {
 		fmt.Println(err)
 		os.Exit(1)
 	}
-
 }